[ROS-69] social cards bug (#211)
* [ROS-69] Testing this theory

* ok, the same theory, but only without `revalidate`

* Looks like it's been fixed

* Test coverage for the fix
