+++ components/Cards/Comment.tsx
@@ -2,6 +2,11 @@ import { PostComment } from '../..';
 import { UserProfile, useUser } from '@auth0/nextjs-auth0';
 import { useEffect, useState } from 'react';
 import fetch from 'isomorphic-fetch';
+import Image from 'next/image';
+import defaultAvatar from '../../public/defaultAvatar.png';
+import { shimmer, toBase64 } from '@components/Utils/MDXComponents';
+
+const testEnv = process.env.NODE_ENV === 'test';
 
 export default function Comment({
   postComment,
@@ -69,10 +74,22 @@ export default function Comment({
           }
         >
           <div className="flex items-center">
-            <img
+            <Image
               className="h-12 w-12 rounded-full"
               alt={`${userProfile?.nickname}'s avatar`}
-              src={userProfile?.picture ? userProfile.picture : ''}
+              src={
+                userProfile?.picture
+                  ? userProfile.picture
+                  : testEnv
+                  ? 'https://rosnovsky.us/static/avatar.png'
+                  : defaultAvatar
+              }
+              placeholder={userProfile?.picture ? 'empty' : 'blur'}
+              blurDataURL={`data:image/svg+xml;base64,${toBase64(
+                shimmer(50, 50)
+              )}}`}
+              width={50}
+              height={50}
             />
             <div className="ml-2">
               <div className="text-sm ">
