+++ components/Cards/Comments.tsx
@@ -1,4 +1,5 @@
 import Link from 'next/link';
+import useSWR from 'swr';
 import { PostComment } from '../..';
 import Comment from './Comment';
 
@@ -11,9 +12,23 @@ export default function Comments({
   postTitle: string;
   comments: PostComment[];
 }) {
-  const sortedComments = comments.sort((a, b) => {
-    return a.published_at > b.published_at ? -1 : 1;
-  });
+  const fetcher = (url) => fetch(url).then((res) => res.json());
+  const url =
+    process.env.NODE_ENV === 'test'
+      ? 'https://rosnovsky.us/api/comments/getComments?id=test'
+      : `/api/comments/getComments?id=${postId}`;
+  const { data, error } = useSWR(url, fetcher, { refreshInterval: 1000 });
+
+  if (error) return <div>failed to load</div>;
+  if (!data) return <div>loading...</div>;
+
+  const sortComments = (comments: PostComment[]) => {
+    return comments.sort((a, b) => {
+      return a.published_at > b.published_at ? -1 : 1;
+    });
+  };
+
+  const sortedComments = data ? sortComments(data) : sortComments(comments);
 
   return (
     <div className="my-10 w-full">
