+++ pages/blog/[slug].tsx
@@ -23,16 +23,6 @@ export default function Blog({
 }) {
   const { user } = useUser();
 
-  const fetcher = (url) => fetch(url).then((res) => res.json());
-  const url =
-    process.env.NODE_ENV === 'test'
-      ? 'https://rosnovsky.us/api/comments/getComments?id=test'
-      : `/api/comments/getComments?id=${frontMatter.slug}`;
-  const { data, error } = useSWR(url, fetcher, { refreshInterval: 1000 });
-
-  if (error) return <div>failed to load</div>;
-  if (!data) return <div>loading...</div>;
-
   return (
     <BlogLayout frontMatter={frontMatter}>
       <MDXRemote
@@ -55,7 +45,7 @@ export default function Blog({
         </span>
       )}
       <Comments
-        comments={data ? data : comments}
+        comments={comments}
         postId={frontMatter.slug}
         postTitle={frontMatter.title}
       />
@@ -86,5 +76,5 @@ export async function getStaticProps({ params }) {
     }
   ).then((res) => res.json());
 
-  return { props: { ...post, comments }, revalidate: 1 };
+  return { props: { ...post, comments } };
 }
