+++ tests/unit/comments.test.tsx
@@ -3,7 +3,7 @@
  */
 import React from 'react';
 
-import { render, screen } from '@testing-library/react';
+import { render, screen, waitFor } from '@testing-library/react';
 import '@testing-library/jest-dom';
 import Comments from '../../components/Cards/Comments';
 import { PostComment } from '../..';
@@ -43,8 +43,11 @@ describe('Comments', () => {
     );
     screen.debug();
 
-    expect(screen.getByText('Test comment 2')).toBeInTheDocument();
-  });
+    await waitFor(() => screen.getByText('Comment'));
+
+    });
+  })
+
 
   // Gotta be logged in for these... HOW?!
 
@@ -80,4 +83,3 @@ describe('Comments', () => {
 
   //   expect(screen.getByText('Test comment 2')).toBeInTheDocument();
   // });
-});
